Turn on nodes/pivots numbers; Explicit conversion from int to double to avoid gcc warnings
git-svn-id: https://ferma.svn.sf.net/svnroot/ferma/trunk@268 d51c77f7-620d-0410-a109-9a2f07eaa567

+++ FermaNext/src/TrussUnit/TrussUnit.cpp
@@ -179,8 +179,8 @@ void TrussUnit::paintLoad ( TrussLoad& load, QPoint tailPos, ren_dynarow& baseRe
     else if ( y > 0 )
         forceDirection.setY ( 1 );
 
-    QPoint headPos ( tailPos.x() + 20 * x1 * forceDirection.x(), 
-                     tailPos.y() - 20 * y1 * forceDirection.y() );
+    QPoint headPos ( int(tailPos.x() + 20 * x1 * forceDirection.x()), 
+                     int(tailPos.y() - 20 * y1 * forceDirection.y()) );
     
     solidRenderer solidRend ( baseRend );
     scanline_rasterizer   ras;
